Update to db_lookup to prevent problems with 64bit windows since 'long int' is not the same size as a pointer.
svn:revision:54478
svn:branch:trunk
svn:account:r_weiss

+++ src/librt/db_lookup.c
@@ -87,7 +87,7 @@ int
 db_dirhash(const char *str)
 {
     const unsigned char *s = (unsigned char *)str;
-    long sum = 0;
+    ptrdiff_t sum = 0;
     int i = 1;
 
     /* sanity */
@@ -96,7 +96,7 @@ db_dirhash(const char *str)
 
     /* BSD name hashing starts i=0, discarding first char.  why? */
     while(*s)
-	sum += (long)*s++ * i++;
+	sum += (ptrdiff_t)*s++ * i++;
 
     return RT_DBHASH(sum);
 }
